SCRD-724 ensure that scripts are run from the correct directory
Ensure that run-in-deployer.sh, update-venv.sh, and
test-project-stack.sh are run from a subdirectory of
ardana-dev-tools/ardana-vagrant-models which contains a Vagrantfile, because
this Vagrantfile is required in order to access the deployer node to
which the updated venv will be deployed, and/or to generate an ssh
config file.

In update-venv.sh, do it in such a way that -h / --help still works
from anywhere, and the ARDANA_VERSION calculation is done after the
check, otherwise it could fail when invoked from the wrong directory.

Change-Id: I14ee1c27fd154254d48c5ed7a7f6c1fe4705bf6e