SCRD-1451 Skip media check for RHEL7.2 install
The default RHEL7.2 install target checks the integrity of the ISO
image before using it to install. Since we have already validated
the ISO integrity before running packer, this step is unnecessary,
and wastes time.

This can be avoided by sending an "Up Arrow" keypress as part of
kicking off the boot action.

Change-Id: I26694ec9544457473f3f07539c8e41cdfd9a93f0