SCRD-1659 Manage venv build ansible forks
The current venv build process unconditionally sets the Ansible fork
factor to 100 when kicking off the source retrieval and wheel build
phases on the Build VMs, meaning that ansible will create up to 100
parallel host actions if that many hosts are available; when running
with 3 Build VMs, one for each supported platform, as is the case in
a CI mode run for our standard integration test models, we can have
approximately 150 dynamically created virtual hosts spread across the
three build VMs, which are also being created with nproc - 2 vCPUs,
massively overcommit the system CPU resources. Running so many of
these network retrieval or wheel build actions simultaneous on the
overcommitted CPU resources leads to Ansible SSH timeouts waiting
for these tasks to complete.

To overcome these issues we need to avoid excessively overcommitting
the system CPU resources, and reduce the Ansible fork factor to a
level that is sensible, and pretty much guaranteed to avoid seeing
issues during a CI run, even an initial "cold" run that has to pull
and build all sources from scratch.

To allow for the fact that developers may have more capable systems,
and be more tolerant of the odd failure, if developer mode is enabled
via the ARDANA_DEVELOPER env var setting, we will use a fork factor
of nproc * 3.

Also tweaked the logic of the NO_STOP/vagrant destroy invocation so
as to avoid a "failed" command status causing a "failed" status from
the venv build script if NO_STOP is set.

Change-Id: Ide1623655e488c580b30763c6666ca08881e13e6