SCRD-1975 No hLinux venvs in canary product tarball
We want to ensure that the canary build process for Ardana doesn't pull
in any hLinux venvs to ensure that the tarball size remains acceptable.

As part of getting this working we rename the build-hlinux-artifacts
script to build-distro-artifacts since it in fact builds the required
artifacts for all distros, not just hlinux.

We also add --hlinux option support to build-venv.sh and update-venv.sh
and ensure that if no distro is selected then we default to using hLinux
for now; this will change to SLES soon.

In astack.sh we check for no distro being selected and default to hLinux
for now; this will change to SLES soon. Also simplified check for
whether hLinux can be disabled, i.e. when deployer, controllers and
computes are all SLES or RHEL, we don't need hLinux.

Don't export the PS4 variable in libci.sh; this had side effects that
"broke" some ansible shell actions that run with sh as their shell.

Rework the scripting in update-venv.sh somewhat to simplify managing
multiple distros.

Added/updated TODO statements in various places where we will want to
make changes to make SLES the default when ready.

Change-Id: I82df405e81c00d4fa3ce9a936ecc0a38e1dd5620