SCRD-1370 Don't build hlinux box unnecessarily
The image-build-vagrant-box playbook was missing a when:, so it was
unconditionally building an hlinux box with packer even when
ARDANA_HLINUX_ARTIFACTS was unset. There's no reason other models like
deployerincloud need an hlinux build box when deploying a SLES cloud.
Avoiding unnecessarily building this box saves a great deal of time.
This patch corrects the playbook so that hlinux artifacts are not
unnecessarily built.

Change-Id: Ie99c9c3a8283b2f19ed99a72a4761ba5a49c6b36