SCRD-2398 Setup ardana runtime environment
Add the ardana-env tool whose output can be eval's to set up the runtime
environment for ardana-dev-tools tooling, and update all the tools in
the ardana-dev-tools/bin directory to use this new script.

The ardana-env tool currently will create an ansible venv under the
ardana cache, and output the necessary PATH update to ensure that the
venv's ansible commands are used.

Updated README.md to document the new ardana-env command, and updated
doc/troubleshooting.md to include Nicolas Bock's suggestion to be
explicit about which python version the virtualenv should be based on.

Change-Id: I77de8eb67e2a9ac31f037fce19720857ac07679f