SCRD-2019 Use the packages from the Cloud 8 media
ardana-dev-tools should use the packages from the repos backing
the Cloud8 up in the product rather than some random collection
of build service packages. If there is a reason that a package
is missing then it needs to be added to the product.

Misc updates to the tarball generation process to ensure that we
are using the correct zypper repos appropriately, by running the
zypper addrepo command directly, rather than via the ansible
zypper_repository module.

Ensure that we setup the DHCP provided default route on the SLES
build VM early in the build process so that it is available for
use by the SLES Extras tarball build process. As a result we no
longer need the changes previously added to venv-build.yml for
SCRD-2073.

Update the SLES Extras tarball reference details to point to the
new version which will be uploaded to the Ardana artifacts server
in the near future.

Depends-On: I928d5abb0478a33fdd4db39284cec9cbe05345bf
Depends-On: Idd844f3ef053a84f5eb9fbca2e5f4873850c998d
Depends-On: Ia0b2ef76b9fb333b899a97f2dbb40abca46dc710
Change-Id: I402d2406cb674b6b3743239051cba824d78eb52d