SCRD-2636 Create user specific qcow2 images
Currently when you switch between CI and now-CI mode you have to
rebuild the qcow2 image to change the user account built into the
qcow2 images.

Since we don't have access to the virt-customize tool on the Trusty
based legacy CI systems - the version of libguestfs-tools is too old,
we can't easily edit the qcow2 boxes. An alternative approach is to
maintain separate images, per user account used, which can be used
to create our Vagrant boxes as needed when we switch modes. This
will also allow us to add new modes, such as a Cloud8 one, where
the user account is ardana, with a home directory of /var/lib/ardana.

Also fix the cleanup.sh script used for SLES Vagrant box creation to
delete all registered zypper repositories; the previous version was
stale and was just trying to delete the SLES12 SP2 ISO repo.

Change-Id: Ie03f64cb8bfa9b1e01a5f4ea1ffe5c44564412cd