SCRD-2696 Install systemd service files into /etc/systemd
/usr/lib/systemd is owned by packages so use the correct place to
install custom service files.

Also do not create an Alias with a duplicate name.
The [install] section of a systemd .service file should not create an
Alias= entry with the same name as the .service file itself. This
information is redundant and  will lead to eg:

$ systemctl reenable swift-proxy-server
Failed to execute operation: Invalid argument

when trying to (re)enable a service.

Note: This only happens since the custom .service files moved from
/usr/lib/systemd/system to /etc/system/system .

Also make the copyright check happy.

Depends-on: I66ea91ced1804fefbaa4b995b5fbf42ef42fae16
Change-Id: I55f3895abb0c5103436a99a3e459402d557404e8