SCRD-2398 Revert unnecessary use of ardana-env
ardana-git-grep is different to the other scripts in bin/ in that all
it does is grep sources, and consequently doesn't need any of the
normal bootstrapping.  In fact it doesn't even need to be run in any
kind of Ardana environment.  So revert the dependency on ardana-env
introduced by 93088d0c which causes potentially unwanted installation
of virtualenvs and other components.

Change-Id: Ibea8a94bbc04f4eac95aea88b1dcafa901f87734