SCRD-2083 Enable multi-architecture venvs
Use a new 'package_venv_locations_multi_arch' data structure
to enable venv packages to be installed for multiple
architectures simultaneously when built as RPMs.

For legacy builds, this change should have no effect,
as the only supported architecture is assumed to be
that of the deployer - x86_64 - and the venv installation
paths are kept unchanged.

Change-Id: I5ea2ebe7d0f428511284a0e74f1bb9755709ed96