SCRD-2694 Synchronize the index with current venvs
With the SUSEification effort we deploy the venvs inside an RPM
package, so when a new version of the venv is installed the old
one get removed from the `ardana_packager` directory.

As of today there is not co-installation for multiple versions
of the same venv service.

This patch makes sure that when an old venv tarball is removed,
it is also removed from the venv index.

Change-Id: I4e5cd757657124b2ffb8a809d299183442ddc5ab