Import octavia image fails due to file format (bsc#1083335)
Import octavia image fails due to the file format.
The original guest image file is a tar file, since it used to
have multiple guest images one for ovsvapp and the other for
the amphora.
Right now we only have one guest image which is a qcow2 image.
So let us go ahead and add the image directly with the qcow2
extension.

Change-Id: Icb068a8ee4189de4e2913101464a01c1176c16c4