SCRD-2855 Enable console logs in C8 mode
In legacy mode if you run with the --ci option, you capture the console
logs from the vagrant VMs under the logs/console directory; add similar
support when running in Cloud8 deployment mode.

Note that this can be useful if you want to monitor the VMs for reports
of OOM killer activity during a deployment.

Change-Id: I071b73be1b0c1393baa343130f652a1a99c63bf4