SCRD-2889 Proxy cache RPM files only
The intent of using a caching proxy service on deployer for upstream
Zypper repo access is to accelerate accesses to RPMs such that we only
retrieve them once, no matter how many nodes are configured in a cloud
deployment.

To achieve this we configure the caching proxy (squid) to explicitly
only cache RPM files.

Instead of implementing this as yet another lineinfile edit action, we
switch to using a template to generate the squid.conf.j2 file entirely
and remove the old lineinfile based editing code.  This is due to the
fact that the complexity of the changes we are making via lineinfile
getting to be too much.

Added copyright checking ignorance for the squid.conf template.

Change-Id: Ic47122c55bfbf7fd43f725ca783684dde55e4ec7