SCRD-2897 Use pattern matching for Cloud8 ISO name
Instead of using a hardcoded name, with a specific build id, instead
we use a pattern to identify the candidate ISOs in the SHA256SUMS file
and then download the latest version found.

This helps isolate us from having to submit patches to match when the
target ISO keeps getting updated.

This is also necessitated by the fact that the SHA256SUM file doesn't
usually have a valid entry for the "unversioned" ISO published in the
same place.

Change-Id: I3047e7342ab09abfe6d8796fcd15820e2b4e48ad