SCRD-2907 Cloud8 deployer doesn't need SDK ISO
A Cloud8 deployer doesn't need to attach the SLES SDK ISO, and serve
it out as zypper repo, so remove the existing support for doing so.
However we don't want to remove support for attaching this ISO to the
SLES build VM, or the legacy deployer.

Some re-working of the ardana_vagrant_helper.rb setup_iso() routine
was required to allow us to determine if we are setting up a build
VM or a cloud VM; we now pass the name of the VM to this routine to
allow it to check for "build" in the name.

Change-Id: I2d4c72711b3797fe204bebb5286e9deeac24c4d1