SCRD-2019 Use the SUSE builds of the packages from cloud8
Update SLES Extras tarball generation to use DC8 repos in general as
sources for packages.

Note that liberasurecode-devel is needed only during the venv build
process, so even though the SLES extras tarball contains a separate
SUSE-SLE-12-SP3-Update-Products-Cloud8-standard repo for it, and any
dependendencies that it may have, we don't need that repo setup on
the Cloud nodes for a successful deployment.  This means that there
is no need for a dependent change in osconfig-ansible's legacy SLES
extras repos list.

Added support for the new ARDANA_SLES_NET_REPOS_SRPMS environment
variable, defaulting to false, which determines if we include the
SRPMs in the SLES extras tarball.  This was required in the legacy
product releases, but is no longer necessary.

Finally we enable building the SLES Extras tarball dynamically by
explicitly setting ARDANA_SLES_NET_REPOS to true in the astack.sh
runtime environment, if not already set.  Additionally we change
the ansible check for it's value to default to true if not set,
so that a manually started SLES build VM is setup appropriately.

Change-Id: Ief048b96115026103a9f44680a49daa1e61a7dd5