SCRD-2956 Switch to using rsync for Cloud8 repo mirroring
Replace the existing wget based mirroring process with an rsync based
scheme which has the benefit of properly managing the local mirror by
ensuring that stale files are removed.

Using rsync directly as the ansible synchronize module internally is
based on setting up SSH connections as the transport medium.

Updated the mirroring process to only mirror the enabled repos, based
upon the Cloud8 flavour (HOS or SOC) and specified set of repos to be
used (DC8, DC8S or Pool + Updates) by the astack.sh run.

Additionally fixes the default enabled repos settings to ensure that
DC8S is the only repo used.

Change-Id: Idf0eccc270d1ecfce484a2ed8105c4a2519d528f