SCRD-2968 Exclude ansible localhost from sources sync
We need to ensure that we exclude the ansible hosts/localhost file
when synchronising local ansible sources to the deployer, otherwise
we may overwrite the Cloud8 customisations to that file which are
used to determine whether we are running in legacy mode or not.

Also exclude the LICENSE and .copyrightignore files.

Change-Id: Ieee622737484143c5a29a94a4e370dd3ce654207