SCRD-3000 Support for specifying tempest filter
Allow the user to specify the tempest testing filter to use when we
run the tempest tests after deploying the cloud.

We add the new --run-tests-filter option to implement this, which
implicity selects --run-tests, so you only need to specify one or
the other.

The default filter used is the 'ci' filter.

Updates README.md to reflect new option.

Change-Id: I44a27d566fd82bca4b50a4322eb5d048cd1e0f7a