SCRD-3021 Persistently cache mirros Cloud8 repos
The existing Cloud8 repo mirroring using the ephemeral cache area under
the ardana-dev-tools scratch area, which is deleted after each CI run,
so the repo must be fully mirrored on each run.

Instead we can move the mirror cache to the ~/.cache-ardana area where
it can be persistently cached.

As a result the ardana-scratch role can be removed.

Change-Id: Ifb69303acc224e1856fd4d896fdb9665a10bfcb9