SCRD-3020 Re-work --run-tests handling for Cloud8
Add a new --c8-qa-tests option to trigger building and uploading of QA
venvs in Cloud8 deployments.  This means that if just --run-tests is
specified we will not longer attempt to build the QA venvs, and will
just run the tempest tests against a Cloud8 deployment, which should
save 5-15 mins on the overall time taken to run just tempest tests.

Implicitly enable Cloud8 deployment mode if any of the Cloud8 options
are specified.

Ensure that if someone specifies --ci with --c8 that we still use the
correct account name in run-in-deployer.sh

Change-Id: I34e8b24978c95da94834216f24eb2e7697875627