Fix setup of proxied repos on non-deployer nodes (bsc#1089716)
The check to see if the deployer was enabled to deploy/configure the
proxy cache only worked for the deployer node, and not on other nodes
so update the check to specifically check the deployer node settings
so that all nodes get the same result.

Ensure that we open the caching proxy port (3128) in the firewall
so that nodes can access the proxy server once the osconfig stage
of the deployment has configured the firewall; we achieve this by
adding an additional YAML file with the required rules.

Enhance the ardana-vagrant-ansible helper script so that if it is
run from a <cloud>-vagrant dir, with a Cloud8 deployment, with no
arguments, then it will run the cloud8-setup.yml playbook to update
the cloud environment.

Updated README.md entry for ardana-vagrant-ansible to reflect the
changes made.

Change-Id: I90685ffde2778b76904c5bcc0f649bfa3f8bba62