SCRD-3042 Use specified Cloud8 flavour ISO
Up to now we have always been using the HOS8 flavour ISO even when the
--c8-soc option is specified; this patch re-factors the downloading of
then Cloud8 ISO, and the configuration of it as a zypper repo within
the deployed cloud, so that we use the ISO associated with the flavour
specified.

Moved the suffix for the set of ISOs being downloaded to be specified
by the cloud8.config.iso_type setting, so that we only have to update
it once when moving to a new RC#, GMC or GM set.

Change-Id: I0e9cb41336d317a96a6419f82dec4639b1570b27