SCRD-3129 Improved deployer_address determination process
When running the cloud8-setup.yml against a deployed cloud additional
IP addresses may have been defined for the deployer node and entry 0
of the list of available addresses may not be the correct address any
more.

Instead use the address associated with the deployer's eth2 device
during initial setup, and the address associated with hed2 after it
has been renamed by the osconfig-ansible run.

Additionally added a check that means we will fail out if we can't
determine the deployer address from these two sources, and the user
hasn't provided one.

Change-Id: I9b7f4e7d7a360738532ac2a17445d81224f94adf