SCRD-3128 run-cobbler Cloud8 support
Update the run-cobbler tool to support Cloud8, and make Cloud8 mode
the default.

Add options to explicitly control which mode of operation should be
used, as follows:
  * --legacy
    Use setting equivalent to the old legacy mode of operation,
    e.g. user is 'stack' homed under /home/stack
  * --legacy-ci
    Use setting equivalent to the old legacy CI mode of operation,
    e.g. user is 'ardanauser' homed under /home/ardanauser
  * --c8
    Use setting equivalent to the new Cloud8 mode of operation,
    e.g. user is 'ardana' homed under /var/lib/ardana

Ensure we set the ARDANAUSER & ARDANA_USER_HOME_BASE directories
appropriately depending on which mode we are running in.

Re-work the options handling in run-cobbler similarly to what was
done for astack.sh option handling to make it easier to maintain
going forward.

Change-Id: I11628957553f906da41f6153289110a96ea0714d