SCRD-3139 ardana-env installs appropriate osc package
Add support to ardana-env for installing an appropriate osc package
for use building Cloud8 RPMs. On Ubuntu systems we add an APT repo
entry to the /etc/apt/sources.list.d directory, targetted at the
OBS repos providing osc for (x)Ubuntu.

For now we require that you must have at least version 0.152 of the
osc command.

Additionally we will need the createrepo command to be installed to
initialise the NEW_RPMS directory with appropriate rpm-md metadata.

Ensure that we run apt-get in non-interactive mode.

Change-Id: I68302d98893b86e2a24999106aa63cc5d7166c8a