SCPL-339 Add override repos to Cloud8 deployer and build update RPMS
Define a new set of override repos for the Cloud8 deployer setup,
with one entry pointing at ardana-dev-tools/../NEW_RPMS, which will
be rsync'd to the deployer and added as a zypper repo with priority
98.

Ensure we auto accept zypper GPG keys when running zypper refresh.

NOTE: You will need to manually install the createrepo package and
manually run it against the NEW_RPMS directory before testing this
patch.

The tests for the ardana/next branch are currently checking
out any repos where the sha1 does not match the sha1 in the published
sources.

This adds code to ardana-dev-tools to create RPMs for any repos
that are checkout at the same directory level as ardana-dev-tools
and that we can determine are used when building an RPM.

Change-Id: I6cc3ec33921e458d082b257b24a4cd7170c7124e