Adding monascaclient centos rpm for monasca ansible modules (bsc#1096583)
When nova-deploy is ran against only RHEL compute nodes by limiting
the hosts, monasca ansible module fails as they require monsacaclient.
Following are examples of monsaca ansible modules needing this client
  - monasca_notification_method
  - monasca_alarm_definition

Earlier this issue was not seen as ansible steps executing these modules
are executed only once and SLES cp and compute nodes have this client
installed. When play is limited to RHEL computes, then this issue
appears.

So including related package in RHEL setup for installation.

Still need to identify where/how this package is installed on SLES
compute host and if that step is done for RHEL or not. Earlier failure
can be attributed to fact that monsacclient rpm was not available and
install fail silently on RHEL compute host. Need to check that theory.

Change-Id: If8936aaeddc733148238f164f3929670c2865a32