SCRD-3706 Default to ardana as user
The legacy mode of have the user be stack in DevEnv and ardanauser in
CI is no longer relevant since our CI testing is now Cloud8 based,
which always explicitly sets user and home directory base settings
to ardana and /var/lib respectively.

However many of the tools still default to using the legacy settings
which causes problem when using them in Cloud8 CI testing. To fix this
we update the tools to default to ardana as the user, but supporting
overriding these settings where appropriate using pre-existing env
var settings.

Change-Id: I077459773712a3edbff989cdba6a0233894ccdb6