SCRD-3748 Mirror SLES12 SP3 Pool and Update repos
Ensure that we setup the SLES12SP3 Pool and Update repos in a Cloud8
deployment by mirroring the required repos from the provo-clouddata
mirror site using rsync, and sync them up to the deployer.

Note that this represents an approx 10G increase in disk space usage
on the DevEnv/CI host under ~/.cache-ardana, and also an additional
~10G of disk space consumed on the deployer node.

NOTES:
A better solution might be to ensure we always configure the caching
proxy on the deployer, and always configure these repos to use the
caching proxy.

Change-Id: I03addb911ddc89c074b2e925b31e5b4722ff960e