SCRD-3752 Use version specific vagrant plugins
By setting the VAGRANT_HOME env var to point to a version specific
"home" directory we can avoid some of the issues seen by developers
when the switch between Vagrant versions, e.g. correct plugins are
found, but built for the wrong version of Vagrant. This is done by
the bin/ardana-env script now when it runs.

Convert the vagrant installation and validation process to use a
new dev-env-vagrant role, organised as a true Ansible role to be
more Ansible 2.x compatible. Part of this is the separation of
plugin management to be on a per Vagrant "series" basis, i.e.
Vagrant 1.7.x can have different plugins specified that Vagrant
1.8.x, or Vagrant 1.9, or Vagrant 2.0, and so on. Combined with
the Vagrant version specific VAGRANT_HOME areas that ardana-env
now sets up for us, this will make it much easier to add support
for newer vagrant versions as we identify lists of compatible
plugins.

The new dev-env-vagrant modules should also correctly handle the
case of installing new and updating existing plugins if the list
of plugins for a given Vagrant series is changed; it will not
currently automatically remove any plugins to allow the user to
manually add additional plugins if so desired. It should also
correctly handle the case of detecting if plugins have been built
with the wrong version of Vagrant, automatically installing the
correctly versions.

To ensure that just your Vagrant environment is up to date you
can run the ansible/dev-env-vagrant-install.yml playbook, e.g.

    % eval $(bin/ardana-env)  # setup Ardana runtime env vars
    % ansible-playbook -i ansible/hosts/localhost \
      ansible/dev-env-vagrant-install.yml

Added support for caching downloaded Vagrant version packages
under the ~/.cache-ardana area, in the vagrant/downloads area.
We also retrieve the packages directly from the Hashicorp site
rather than relying on the internal Ardana artifacts server.

Re-ordered the plugin lists for the Vagrant 1.7 stream to avoid
build issues seen in recent CI test runs of this patch.

Setup minimised plugin lists for the Vagrant 1.8, 1.9, 2.0 and
2.1 stream, specifying only those plugins we are actively using
and updating the vagrant-libvirt plugin to the current latest
version (0.0.43), including implementing a patch to re-enable
use of virtio-scsi based SCSI controller in generate domain
definition XML.
NOTE: Additional work required to re-add support for serial
device redirection to files as part of the vagrant-libvirt
0.0.43 patch.

Updated bin/ardana-env and the DevEnv/CI package lists with a
first pass at support for both openSUSE (and SLE) 15 and Ubuntu
18.04 (Bionic). Also added support to bin/ardana-env to detect
and install some additional minimum requirement tools, e.g.
rsync, wget, git; this helps with getting the DevEnv working
on a minimally installed system with openSUSE Leap 42.3 or
Ubuntu Xenial (16.04) or Trusty (14.04). Added the acl package
to the Debian (Ubuntu) package lists as it is being used by the
ansible/dev-env-validate.yml playbook tasks, but wasn't always
being installed as a dependency of other packages.

Updated README.md to reflect impact of changes made to the
Vagrant version support and to indicate which distributions
are supported.

NOTES:
  * Vagrant 1.9 and later streams won't work for RHEL compute
    based setups due to an issue in the vagrant network setup
    code which is assigning the wrong IP to the management
    (eth2) device on the RHEL computes, breaking the serving
    of repos to those nodes.

Change-Id: Ia2d883b4b693ef1a7ad88338363343596d4f9c59