SCRD-3914 Cleanup ansible-lint tooling
Cleanup our ansible-lint tooling, used by that ardana-ansible-syntax
job, so that it doesn't depend on any legacy HOS/Ardana resources,
and is more ansible 2+ friendly.

Update the lint role to be an actual ansible role, since the existing
parttern can be directly converted to an true ansible role.

Ensure we install appropriate versions of the relevant PyYAML, ansible
and ansible-lint packages via a requirements.txt file, found in the
ansible/roles/lint/files directory. This is required as the current
latest available version of PyYAML (4.1) breaks ansible lint.

Ensure we use the python from the lint venv when running the python
based scripts for sanity checking the flattened Ardana ansible code
base, otheerwise we may end up using the system python command which
may be Python3 based on newer systems.

Change-Id: Ibb095193197af87312a6c5cbb96dc343c426b8b7