SCRD-3969 Add support for openSUSE Leap 15
Fixed the connect_deployer_to_extnet.yml task to not depend on the
ifconfig command which is not available in openSUSE Leap 15.

Also re-worked Vagrant dependent package install process to always run
even if vagrant is already installed, to ensure that we install any
dependencies that are indirectly required, e.g. to be able to build
plugins.

Added support for post-Vargrant install workarounds, currently needed
by the SUSE family to deal with broken libreadline.so.X libraries in
the embedded Vagrant tree. This applies to both Leap 42.3 (Vagrant
1.9.x) and Leap 15.0 (Vagrant 2.x) variants.

Change-Id: I36b57daf8393596e27a8e2251fe0763ded86b073