SCRD-4015 Add dev-env-setup helper script
Add a new helper script, dev-env-setup, which will endeavour to set up
the users system environment to be ready to run the Ardana DevEnv via
the astack.sh script.

This is equivalent to what the astack.sh command does during the setup
phase, which can be skipped via the --no-setup option if you have run
it successfully before.

The output of this tool can be evaluated from the shell command line,
similar to the ardana-env command, setting up environment variables
appropriately to be able to run astask.sh script.

The script also allows you specify the desired versions of Ansible
and Vagrant to be installed, and will setup the appropriate env vars
to ensure that subsequent runs of ardana-env and astack.sh use those
versions.

There is also a --clean option which will clean out any existing
Ansible venvs and Vagrant home areas for the relevant versions, and
force a clean re-install of those tools.

The --vagrant-only option can be used to run the Vagrant specific
DevEnv setup playbook, dev-env-vagrant-install.yml, instead of the
full setup playbook, dev-env-install.yml.

The --env-only option can be used to just set the appropriate env
vars without triggering the setup of the Ardana DevEnv.

To support the dev-env-setup script the ardana-env script has been
enhanced to take a new --clean option telling it to clean out the
existing Ansible virtualenv and Vagrant home areas. Additionally
the ARDANA_ANSIBLE_VERSION env var can be used to specify which
version of Ansible we want ardana-env to install.

Updates to the README.md to reference the new dev-env-setup tool,
and correct some minor errors and formatting/layout tweaks.

Updated the --pre-destroy option's usage description in astack.sh
to synchronise it's text with that of the README.md file.

Change-Id: Iaf1b328ee120e13d1b3f0bcbacd712ecb00a41e7