SCRD-4144 Setup ISO links in Cloud8 deployer
The Cloud8 deployer has the necessary ISOs attached to it as CDROM
devices, so there is no need to attempt to upload the ISO images to
the deployer. We just need to ensure we create the appropriate links
in the deployer account for the cobbler-deploy.yml playbook to find
the ISOs.

Change-Id: Ic0b1c4c46c7bb0288c68bd0d8dc1df2968218ab3