SCRD-4199 Use PyPI directly rather than bandersnatch
Switch to using PyPI directly, rather than that the simple index of
the local bandersnatch server, when creating the ardana-dev-tools
ansible venv.

Recent CI stability isses associated with creating the ansible venv
may be related to upstream changes to finish implementing/deploying
PyPI 2.0, that our bandersnatch server may not be interacting with
correctly.

Change-Id: I317dfe9a4e5f2485bc2eded4bbf93597853cc2f0