make mode checking smarter (bsc#1102968)
The way we used to check for mode syntax is very narrow and
inflexible. Therefore, it will chirp if we use a sticky bit
or we don't use a strick syntax for the symbolic notation.
Furthermore, it doesn't seem to support sticky and special
execution bits in either form. This patch is to make the
mode checking a bit more smarter. To accomodate the complete
syntax described here.

https://en.wikipedia.org/wiki/Chmod

Change-Id: Ic8a30283c483eebe09eeb6ccd04d5e8972de4487