SCRD-4810 Update cobbler test tools with SLES12SP4 support
Update run-cobbler with --sles12sp3 and --sles12sp4 options, and to
take --c8-pool and --c9-pool options. Also change the default model
to demo rather than dac-min, as we don't actually deploy the cloud,
and demo has a smaller memory footprint. Finally add support for
passing through the --no-update-rpms option to astack.sh.

Update the add-distros.py script to allow us to specify which version
of SLES to use as the default distro. Also cleaned up the logic for
setting a desired distro for a server so that we only set a distro
for nodes that aren't going to be the default distro.

Update cobbler-deploy.sh to check for sles_version_name attribute
and if found, change it to the specified SLES version, and commit
the change.

Update astack.sh to call cobbler-deploy.sh passing in the active SLES
version. Also implicitly default to SLES12SP4 if deploying a Cloud 9
cloud in general.

Updated README.md to reflect recent updates, including SLES12 SP4 and
Cloud 9 support.

Change-Id: I6ed471b1ebba5a07a4e707bbcf81968ecea815cd