SCRD-4968 Install correct kvm_server package
The name of the kvm_server pattern package differs between openSUSE
and SLES so switch to using the distribution name rather than the
operating system family name to determine the appropriate set of
packages to install for KVM/QEMU support.

Updated Debian and SUSE family package installs to use the correct
list based on the calcualted value of the new kvm_distro fact.

Change-Id: I1092a79e4a328e9130fa260c6006197cbc2acfad