SCRD-5002 Make Cloud9 default cloud version
Updated the astack.sh default settings to make Cloud 9 deployment
the default. Also update deprecated/legacy options to produce an
appropriate error message if specified to indicate they are not
supported any longer.

Updated default milestone for Cloud 9 to M2, latest available at
this time; can be overriden by setting ARDANA_CLOUD9_MILESTONE
to the desired version.

Updated the README to reflect the fact that the legacy deployment
mode is no longer supported, updating or removing entries related
to the legacy deployment process. Also moved the copyright text
to the end of the fike to make the README.md file more readable
in rendered form.

depends-on: I3c50c51253153779534e8a1e120e16c8859113bb

Change-Id: Iec05ab78bedd8c61079cd842d40e261470aab362