SCRD-5116 Vagrant provisioning LVM naming agnostic
Update the Vagrant VM provisiong playbooks that setup the root
file system partition to be agnostic of the names of the LVM
VG and LV used for the root file system. Elimimated the use of
hardcoded device names as well, instead using the information
that ansible dynamically collects from the system that is being
provisioned.

Also updated the code to explicitly check for the tools that we
need, rather than just checking for the presence of the parted
package, which will work better when dealing with images that
have been customised/optimised.

Misc cleanup of this code that will make it easier to port to
Ansible V2+ eventually.

While testing these changes I noticed that the deployer was being
provisioned twice; fixed the provision_deployer() method in the
ardana_vagrant_helper.rb to avoid this.

Depends-On: I904db17c59731685c36bea1459c2cd91ffc2fe57
Change-Id: I99bd14404638752410e3ad43a9492e90ccb6e614