SCRD-5471 Cleanup ISO based repo management
Up to now we have always attached the SLES, Cloud and optionally
RHEL ISOs to the deployer, and have always set up the associated
zypper or yum repos if they are detected.

However we really should only be attaching them if we need them,
and should only configure associated package repos if needed, and
we never actually need to configure the SLES ISO as a package repo
because we have the associated SLES pool repo already available.

We need the ISOs attached to the deployer only when testing:
  * with just the content of the ISO, e.g. testing a milestone release
  * testing cobbler re-imaging of the associated distro, e.g. SLES or
    RHEL, where we need the ISOs available to support import into
    cobbler.

We don't want to configure the ISO based package repos unless:
  * for Cloud9, we are testing a milestone release
  * we are deploying RHEL computes, in which case we need RHEL ISO
    mounted to provide yum repo
    * need all ISOs attached as RHEL ISO is expected to be 3rd ISO

To support explicitly selecting just the ISO as package source new
--c8-iso/--c9-iso options will be needed.

Updated default Cloud9 milestone to M4.

Fixed some typos in run-cobbler help message.

Improved logic around detecting if running in cobbler mode, and
whether we need to enable the optional RHEL repo. Not including it
in the cobbler testing when RHEL computes are enabled should improve
CI testing stability as the process of contructing the optional RHEL
repo is prone to network stalls/breakages.

Add --download-api to osc build arguments to improve package download
performance and avoid testing instability due to network glitches in
general, and strange errors when running on some Ubuntu environments.

Change-Id: I385db370f87d8568385af8763050b6e0ac5b9948