SCRD-5547 Fix Cloud8 mode for astack.sh
The initial work to enable the master branch to support both Cloud 8
and 9 deployments was incomplete, resulting in Cloud 8 deployments
being attempted using Cloud9 input models. To avoid these issues we
must ensure that we use the stable/pike branch when running a Cloud 8
deployment.

Also fix the default cloud version in dev-env-cloud role to be 9 as
that is what astack.sh now defaults to setting up.

Change-Id: Id25378c7620fcb8a8a10468c57b1d4d7505296df