SCRD-4810 Update to SLE_12_SP4 GMC build
Now that the GMC candidate build is available we can switch to using
it for the packer image creation process, and as the ISO to import
for cobbler testing.

NOTE: Updated sles12sp4-ardana.qcow2 images built using the GMC build
will be uploaded to the tarballs server to match this change.

Change-Id: Iaed7ac3594afbf0062894e6a11f06427dad3a748