SCRD-5737 astack.sh IBS/OBS integration
Add support to the astack.sh command to allow a user to specify one
or more IBS or OBS projects or project packages to be pulled into
the cloud deployment. This can be done either by specifying that
a projects repo should be added to the set of zypper repositories
that are configured on the nodes in the cloud, or by downloading
the RPMs for the specified project or project/package to include
them in the NEW_RPMS area (at the same level as ardana-dev-tools
clone) which is used to create the Overrides zypper repo on the
deployer node.

Two sets of new options have been defined, which can be repeated
multiple times to specify as many projects or packages as needed.

Add IBS/OBS projects as zypper repos to to SLES nodes:
======================================================
  --ibs-repo <project>
  --obs-repo <project>
    Both of these options take a project specifier, e.g.
    home:fmccarthy:branches:Cloud:OpenStack:Rocky:venv, and add
    the appropriate zypper repo to the SLES nodes in your astack.sh
    cloud.
    NOTES:
      * You must enable publishing for your project for this to work!
      * This can be quite slow over a VPN connection or if there is
        limited bandwidth between your nodes and the build service.

Include project (or project package) RPMs in NEW_RPMS:
======================================================
  --ibs-prj <project[/package]>
  --obs-prj <project[/package]>
    Both of these options can take a project or a project/package
    specifier, e.g. home:fmccarthy:branches:Devel:Cloud:9:Staging or
    home:fmccarthy:branches:Devel:Cloud:9:Staging/ardana, to specify
    that you either include all the RPMs built for that project, or
    just the RPMs build for that package in the project.

Implemented the helper script, get_buildservice_project_rpms, which
can be used to download RPMs, using the osc getbinaries command, from
specified IBS or OBS projects.

Updated README.md to reflect these changes and add some explanation
about the update_rpm.sh script.

Change-Id: I9dec3e8f5164d69c42f0e5e3fb57283b2e000e65