SCRD-5557 Add support for IPv6 libvirt networks
We need to add IPv6 subnets to support IPv6 in control plane.
In this patch we add IPv6 subnets to existing libvirt networks,
as pure IPv6 subnets are not yet supported by the plugin. For this,
we move to vagrant-libvirt plugin version 0.0.45 for legacy vagrant.

Also added different IP addressing options for Ardana CP. We now
have '--ipv4', '--ipv6' and '--ipv6-all' options for astack.

Change-Id: Id715a7cee77359cb1c90fe9699999abe2a947086