SCRD-5802 Add retries for rsync and ISO retrieval
Add retries to the get_url actions used to retrieve ISO files and to
the command actions used to run rsyncs of the zypper repos to deal
with some of the known network glitches that can intermittenly affect
these operations.

Note that we still will see failures if SHA256SUMS files associated
with sets of ISOs in a Provo IBS or Clouddata mirror are updated to
reference ISO files that haven't yet been mirrored to Provo yet, as
has happened with the SLE 12 SP4 GMC candidate ISOs this last week.

Change-Id: I7bdd38c853b13e8462d158d773a9065aaa16150a